Wzorzec bridge 3.1-3.2 Stworzenie klasy IntArrayStack i reimplementacja klasy DefaultCountingOutRhymer.
