+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -8,8 +8,7 @@ public class HanoiRhymer extends DefaultCountingOutRhymer {
         return totalRejected;
     }
 
-    @Override
-    protected void countIn(int in) {
+    @Override public void countIn(int in) {
         if (!callCheck() && in > peekaboo())
             totalRejected++;
         else
